Unify and put GH access token to the lambda_shared
