+++ tests/ci/cancel_and_rerun_workflow_lambda/lambda_shared
@@ -1,0 +1,1 @@
+../lambda_shared_package/lambda_shared
