+++ tests/ci/runner_token_rotation_lambda/requirements.txt
@@ -1,3 +1,1 @@
-requests<2.30
-PyJWT
-cryptography<38
+../lambda_shared_package[token]
