+++ tests/ci/terminate_runner_lambda/app.py
@@ -5,86 +5,12 @@ import json
 import sys
 import time
 from dataclasses import dataclass
-from typing import Any, Dict, List, Tuple
+from typing import Any, Dict, List
 
 import boto3  # type: ignore
-import requests  # type: ignore
-import jwt
 
 from lambda_shared import RunnerDescriptions, list_runners
-
-
-def get_key_and_app_from_aws() -> Tuple[str, int]:
-    secret_name = "clickhouse_github_secret_key"
-    session = boto3.session.Session()
-    client = session.client(
-        service_name="secretsmanager",
-    )
-    get_secret_value_response = client.get_secret_value(SecretId=secret_name)
-    data = json.loads(get_secret_value_response["SecretString"])
-    return data["clickhouse-app-key"], int(data["clickhouse-app-id"])
-
-
-def get_installation_id(jwt_token: str) -> int:
-    headers = {
-        "Authorization": f"Bearer {jwt_token}",
-        "Accept": "application/vnd.github.v3+json",
-    }
-    response = requests.get("https://api.github.com/app/installations", headers=headers)
-    response.raise_for_status()
-    data = response.json()
-    for installation in data:
-        if installation["account"]["login"] == "ClickHouse":
-            installation_id = installation["id"]
-            break
-
-    return installation_id  # type: ignore
-
-
-def get_access_token(jwt_token: str, installation_id: int) -> str:
-    headers = {
-        "Authorization": f"Bearer {jwt_token}",
-        "Accept": "application/vnd.github.v3+json",
-    }
-    response = requests.post(
-        f"https://api.github.com/app/installations/{installation_id}/access_tokens",
-        headers=headers,
-    )
-    response.raise_for_status()
-    data = response.json()
-    return data["token"]  # type: ignore
-
-
-@dataclass
-class CachedToken:
-    time: int
-    value: str
-    updating: bool = False
-
-
-cached_token = CachedToken(0, "")
-
-
-def get_cached_access_token() -> str:
-    if time.time() - 550 < cached_token.time or cached_token.updating:
-        return cached_token.value
-    # Indicate that the value is updating now, so the cached value can be
-    # used. The first setting and close-to-ttl are not counted as update
-    if cached_token.time != 0 or time.time() - 590 < cached_token.time:
-        cached_token.updating = True
-    private_key, app_id = get_key_and_app_from_aws()
-    payload = {
-        "iat": int(time.time()) - 60,
-        "exp": int(time.time()) + (10 * 60),
-        "iss": app_id,
-    }
-
-    encoded_jwt = jwt.encode(payload, private_key, algorithm="RS256")
-    installation_id = get_installation_id(encoded_jwt)
-    cached_token.time = int(time.time())
-    cached_token.value = get_access_token(encoded_jwt, installation_id)
-    cached_token.updating = False
-    return cached_token.value
+from lambda_shared.token import get_access_token_by_key_app, get_cached_access_token
 
 
 @dataclass
@@ -284,6 +210,8 @@ if __name__ == "__main__":
         with open(args.private_key_path, "r") as key_file:
             private_key = key_file.read()
 
+    token = get_access_token_by_key_app(private_key, args.app_id)
+
     sample_event = {
         "AutoScalingGroupARN": "arn:aws:autoscaling:us-east-1:<account-id>:autoScalingGroup:d4738357-2d40-4038-ae7e-b00ae0227003:autoScalingGroupName/my-asg",
         "AutoScalingGroupName": "my-asg",
@@ -328,14 +256,4 @@ if __name__ == "__main__":
         "Cause": "SCALE_IN",
     }
 
-    payload = {
-        "iat": int(time.time()) - 60,
-        "exp": int(time.time()) + (10 * 60),
-        "iss": args.app_id,
-    }
-
-    encoded_jwt = jwt.encode(payload, private_key, algorithm="RS256")
-    installation_id = get_installation_id(encoded_jwt)
-    access_token = get_access_token(encoded_jwt, args.app_id)
-
-    main(access_token, sample_event)
+    main(token, sample_event)
