+++ tests/ci/terminate_runner_lambda/requirements.txt
@@ -1,3 +1,1 @@
-../lambda_shared_package
-PyJWT
-cryptography<38
+../lambda_shared_package[token]
