+++ tests/ci/workflow_approve_rerun_lambda/requirements.txt
@@ -1,3 +1,1 @@
-requests<2.30
-PyJWT
-cryptography<38
+../lambda_shared_package[token]
